pc - WIP on better jobs testing
One of the challenges is understanding the User and Security context.
We want to ensure that when the job runs, it runs in the security context
of the user that started the job.   This becomes easier if jobs can only be
started by admins, but that's a very limiting restriction.